+++ framework/mesh/UnpartitionedMesh/lua/create.cc
@@ -1,6 +1,6 @@
 #include "chi_lua.h"
 
-#include "mesh/UnpartitionedMesh/chi_unpartitioned_mesh.h"
+#include "mesh/UnpartitionedMesh/unpartitioned_mesh.h"
 #include "chi_runtime.h"
 #include "chi_log.h"
