+++ framework/mesh/UnpartitionedMesh/lua/unpartition_mesh_lua_utils.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "mesh/UnpartitionedMesh/chi_unpartitioned_mesh.h"
+#include "mesh/UnpartitionedMesh/unpartitioned_mesh.h"
 
 #include "chi_lua.h"
